vlm-ui-eval/pyproject.toml
[project]
name = "vlm-ui-eval"
version = "0.1.0"
description = "Test harness for evaluating VLM models on UI design assessment"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    "bitsandbytes>=0.42.0",
    "pillow>=10.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["ruff", "mypy"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
